SmartNexa/requirements.txt
django
djangorestframework
mysqlclient
Pillow
djangorestframework-simplejwt
django-cors-headers